# dev-docs/openapi-mvp.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OSMU MVP API",
    "version": "0.1.0-prototype",
    "description": "Machine-readable contract snapshot for the current OSMU prototype. It covers the REST control plane and the MVP S3-compatible alias surface used by the local release gate."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local prototype backend"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/buckets": {
      "get": {
        "tags": ["Buckets"],
        "operationId": "getBuckets",
        "summary": "List buckets accessible to the user",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      },
      "post": {
        "tags": ["Buckets"],
        "operationId": "createBucket",
        "summary": "Create a bucket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateBucketRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "201": { "$ref": "#/components/responses/JsonOk" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/buckets/{bucketName}": {
      "get": {
        "tags": ["Buckets"],
        "operationId": "getBucket",
        "summary": "Read bucket detail",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "404": { "$ref": "#/components/responses/Error" }
        }
      },
      "delete": {
        "tags": ["Buckets"],
        "operationId": "deleteBucket",
        "summary": "Delete an empty bucket",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/s3/{bucketName}": {
      "put": {
        "tags": ["S3 Compatibility"],
        "operationId": "s3CreateBucket",
        "summary": "S3-style bucket create or bucket tag/lifecycle update",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "security": [
          { "bearerAuth": [] },
          { "osmuAccessKey": [] },
          { "awsSigV4": [] }
        ],
        "responses": {
          "200": { "description": "OK" },
          "201": { "description": "Created" }
        }
      },
      "head": {
        "tags": ["S3 Compatibility"],
        "operationId": "s3HeadBucket",
        "summary": "S3-style bucket HEAD",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "security": [
          { "bearerAuth": [] },
          { "osmuAccessKey": [] },
          { "awsSigV4": [] }
        ],
        "responses": {
          "200": { "description": "Bucket exists" }
        }
      },
      "get": {
        "tags": ["S3 Compatibility"],
        "operationId": "s3ListObjectsOrBucketConfig",
        "summary": "S3-style ListObjects/ListObjectsV2, bucket location, tagging, lifecycle, or multipart upload list",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "security": [
          { "bearerAuth": [] },
          { "osmuAccessKey": [] },
          { "awsSigV4": [] }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/XmlOk" }
        }
      },
      "post": {
        "tags": ["S3 Compatibility"],
        "operationId": "s3DeleteMultipleObjects",
        "summary": "S3-style multi-object delete",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/XmlOnly" },
        "security": [
          { "bearerAuth": [] },
          { "osmuAccessKey": [] },
          { "awsSigV4": [] }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/XmlOk" }
        }
      },
      "delete": {
        "tags": ["S3 Compatibility"],
        "operationId": "s3DeleteBucketOrBucketConfig",
        "summary": "S3-style bucket delete, tag delete, or lifecycle delete",
        "parameters": [
          { "$ref": "#/components/parameters/BucketName" }
        ],
        "security": [
          { "bearerAuth": [] },
          { "osmuAccessKey": [] },
          { "awsSigV4": [] }
        ],
        "responses": {
          "204": { "description": "Deleted" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "osmuAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OSMU-Access-Key"
      },
      "awsSigV4": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AWS Signature Version 4 authorization header or equivalent presigned query parameters."
      }
    },
    "parameters": {
      "BucketName": {
        "name": "bucketName",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9.-]{1,61}[a-z0-9]$" }
      },
...
    },
    "requestBodies": {
...
      "XmlOnly": {
        "required": true,
        "content": {
          "application/xml": {
            "schema": { "type": "string" }
          },
          "text/xml": {
            "schema": { "type": "string" }
          }
        }
      },
...
    },
    "responses": {
      "JsonOk": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "JsonList": {
        "description": "JSON list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ListResponse" }
          }
        }
      },
      "XmlOk": {
        "description": "XML response",
        "content": {
          "application/xml": {
            "schema": { "type": "string" }
          },
          "text/xml": {
            "schema": { "type": "string" }
          }
        }
      },
...
      "Error": {
        "description": "OSMU JSON error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "CreateBucketRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "ownerType": { "type": "string", "enum": ["USER", "ORG"] },
          "ownerId": { "type": "string" },
          "quotaBytes": { "type": "integer", "format": "int64", "minimum": 0 },
          "visibility": { "type": "string", "default": "PRIVATE" }
        }
      },
      "ListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          },
          "nextCursor": { "type": "string", "nullable": true }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "requestId": { "type": "string" }
        }
      }
    }
  }
}